requirements.txt
flask==3.0.3
python-dotenv==1.0.1

langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.0.3
langchain-pinecone==0.1.3

pinecone-client

sentence-transformers==2.7.0
transformers==4.41.2
torch==2.3.1

google-genai==0.3.0

pypdf==4.2.0
tqdm==4.66.4
numpy==1.26.4
faiss-cpu==1.8.0